requirements.txt
facebook-sdk==2.0.0

flask==0.11.1

flask-sqlalchemy==2.1

Flask-PyMongo